add support for additional zip file extensions
This commit allows the user to specify additional zip extensions in the config file.
This is implemented by adding a new optional field to the config file, additional_extensions_zip, which is a list of strings.
If this field is present, the program will use the specified extensions in addition to the default ones.

Internally, the AdapterMeta struct that store file extension matchers is moved into the ZipAdapter struct, and the constructor for ZipAdapter is modified to take the additional extensions as a parameter.
The AdapterMeta struct is then constructed with the default extensions and the additional extensions, if any.